# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Anime Power Battles - Evidence-Based Analysis Engine
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Anime {
  id: number;
  name: string;
  nameAr: string;
  genre: string;
  /** @nullable */
  imageUrl?: string | null;
  characterCount: number;
  /** @nullable */
  description?: string | null;
}

export interface Character {
  id: number;
  name: string;
  nameAr: string;
  animeId: number;
  animeName: string;
  animeNameAr?: string;
  tier: string;
  /** @nullable */
  imageUrl?: string | null;
  /** @nullable */
  powerLevel?: number | null;
  /** @nullable */
  description?: string | null;
  evidenceCount?: number;
}

export type CharacterDetailMetrics = {[key: string]: string};

export interface Evidence {
  id: number;
  characterId: number;
  metric: string;
  content: string;
  sourceType: string;
  /** @nullable */
  seriesName?: string | null;
  /** @nullable */
  chapterEpisode?: string | null;
  /** @nullable */
  pageScene?: string | null;
  confidenceLevel: string;
  isDirect?: boolean;
  createdAt?: string;
}

export interface CharacterDetail {
  id: number;
  name: string;
  nameAr: string;
  animeId: number;
  animeName: string;
  animeNameAr?: string;
  tier: string;
  /** @nullable */
  imageUrl?: string | null;
  /** @nullable */
  powerLevel?: number | null;
  /** @nullable */
  description?: string | null;
  evidence: Evidence[];
  metrics?: CharacterDetailMetrics;
}

export interface BattleRequest {
  character1Id: number;
  character2Id: number;
  /** @nullable */
  context?: string | null;
}

export interface ReasoningStep {
  step: number;
  description: string;
  evidenceIds?: number[];
  conclusion: string;
}

export interface MetricComparison {
  metric: string;
  metricAr: string;
  character1Score: number;
  character2Score: number;
  /** @nullable */
  winner: string | null;
  reasoning: string;
  evidenceIds?: number[];
}

export interface BattleAnalysis {
  id: number;
  character1: Character;
  character2: Character;
  /** @nullable */
  winner: string | null;
  /** @nullable */
  winnerId?: number | null;
  confidenceScore: number;
  confidenceLabel?: string;
  reasoning: string;
  reasoningSteps?: ReasoningStep[];
  metricComparisons?: MetricComparison[];
  evidenceUsed: Evidence[];
  contradictions?: string[];
  limitations?: string[];
  verdict?: string;
  createdAt: string;
}

export interface BattleSummary {
  id: number;
  character1Name: string;
  character1NameAr?: string;
  character2Name: string;
  character2NameAr?: string;
  /** @nullable */
  winner: string | null;
  confidenceScore: number;
  createdAt: string;
}

export interface StatsOverview {
  totalAnime: number;
  totalCharacters: number;
  totalBattles: number;
  totalEvidence: number;
  recentBattles?: BattleSummary[];
  topCharacters?: Character[];
}

export type ListCharactersParams = {
animeId?: number;
search?: string;
};

export type ListBattlesParams = {
limit?: number;
};

export type ListEvidenceParams = {
characterId: number;
metric?: string;
};