# lib/api-zod/src/generated/types/anime.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Anime Power Battles - Evidence-Based Analysis Engine
 * OpenAPI spec version: 0.1.0
 */

export interface Anime {
  id: number;
  name: string;
  nameAr: string;
  genre: string;
  /** @nullable */
  imageUrl?: string | null;
  characterCount: number;
  /** @nullable */
  description?: string | null;
}