# lib/api-zod/src/generated/types/battleAnalysis.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Anime Power Battles - Evidence-Based Analysis Engine
 * OpenAPI spec version: 0.1.0
 */
import type { Character } from './character';
import type { Evidence } from './evidence';
import type { MetricComparison } from './metricComparison';
import type { ReasoningStep } from './reasoningStep';

export interface BattleAnalysis {
  id: number;
  character1: Character;
  character2: Character;
  /** @nullable */
  winner: string | null;
  /** @nullable */
  winnerId?: number | null;
  confidenceScore: number;
  confidenceLabel?: string;
  reasoning: string;
  reasoningSteps?: ReasoningStep[];
  metricComparisons?: MetricComparison[];
  evidenceUsed: Evidence[];
  contradictions?: string[];
  limitations?: string[];
  verdict?: string;
  createdAt: string;
}