# lib/api-zod/src/generated/types/battleRequest.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Anime Power Battles - Evidence-Based Analysis Engine
 * OpenAPI spec version: 0.1.0
 */

export interface BattleRequest {
  character1Id: number;
  character2Id: number;
  /** @nullable */
  context?: string | null;
}