# lib/api-zod/src/generated/types/character.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Anime Power Battles - Evidence-Based Analysis Engine
 * OpenAPI spec version: 0.1.0
 */

export interface Character {
  id: number;
  name: string;
  nameAr: string;
  animeId: number;
  animeName: string;
  animeNameAr?: string;
  tier: string;
  /** @nullable */
  imageUrl?: string | null;
  /** @nullable */
  powerLevel?: number | null;
  /** @nullable */
  description?: string | null;
  evidenceCount?: number;
}