# lib/api-zod/src/generated/types/reasoningStep.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Anime Power Battles - Evidence-Based Analysis Engine
 * OpenAPI spec version: 0.1.0
 */

export interface ReasoningStep {
  step: number;
  description: string;
  evidenceIds?: number[];
  conclusion: string;
}