# lib/api-zod/src/generated/types/statsOverview.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Anime Power Battles - Evidence-Based Analysis Engine
 * OpenAPI spec version: 0.1.0
 */
import type { BattleSummary } from './battleSummary';
import type { Character } from './character';

export interface StatsOverview {
  totalAnime: number;
  totalCharacters: number;
  totalBattles: number;
  totalEvidence: number;
  recentBattles?: BattleSummary[];
  topCharacters?: Character[];
}